pyproject.toml
[project]
name = "card-fraud-platform"
version = "0.1.0"
description = "Shared infrastructure orchestrator for Card Fraud services"
requires-python = ">=3.12"
readme = "README.md"
dependencies = ["pyyaml", "httpx"]

[project.scripts]
platform-up = "scripts.platform_up:main"
platform-down = "scripts.platform_down:main"
platform-status = "scripts.platform_status:main"
platform-reset = "scripts.platform_down:reset"
platform-sync-secrets = "scripts.sync_shared_secrets:main"
platform-sync-configs = "scripts.sync_platform_configs:main"
platform-check = "scripts.platform_check:main"
platformctl = "scripts.platformctl:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
